MoE a16wmix: move SiLU/SiTUv2 activation helpers into act.py
Extract _silu_mul_batch/_sigmoid_f32/_tanh_f32/_situ_mul_batch (and LOG2E) from
gemm1.py into a leaf act.py that uses tensor_shim._to_raw (equivalent to gemm1's
local _raw for these fx inputs, so no circular import). gemm1 imports the two
batch helpers it calls. Pure move; validated logits_diff ~1.5e-5 (a16w4) /
2e-5 (a8w4).
